feat(table): add responsive stacked table option (#1407)
* feat(table): add responsive stacked table option

This adds a new prop `stacked`, which allows the table to render in a visually stacked format based on breakpoint. (CSS assumes standard bootstrap breakpoints).

If no value is passed to `stacked`, the table will always be rendered as stacked.

If `stacked` is set to a breakpoint of `sm`, `md`, `lg`, or `xl`, then the table will only appear stacked at view ports below the specified breakpoint.

When visually stacked, the table thead and tfoot will not be shown.  Cells will be prefixed with the fields label (not the scoped HEAD slot). Because of this, it is not possible to click the "headings" to sort the columns.  Users however, can provide an external control (such as b-form-select) to control the sorting and sort direction (via b-table's `sort-by` and `sort-desc` syncable props).

* Update table.vue

* Update table.vue

* Update table.vue

* Update table.vue

* Update table.vue

* Update table.vue

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* add props for adding classes to thead, tbody and tfoot

new props:
- `thead-class` Class(es) added to `<thead>`
- `thead-tr-class` Class(es) added to  `<tr>` element inside `<thead>`
- `tbody-class` Class(es) added to `<tbody>`
- `tbody-tr-class` Class(es) added to all `<tr>` elements inside `<tbody>`
- `tfoot-class` Class(es) added to `<tfoot>`
- `tbody-tr-class` Class(es) added to `<tr>` element inside `<tfoot>`

* ESLint

* Update README.md

* Create _table.js

* Update table.vue

* ESLint
